requirements.txt
# CUA (Computer Use Agent) Requirements
# Generated from pyproject.toml files and Python imports analysis

# Core dependencies
openai<1.100.0
anthropic>=0.67.0
pydantic>=2.11.1
httpx>=0.27.0
aiohttp>=3.9.3
asyncio
anyio>=4.4.1
typing-extensions>=4.12.2
python-dotenv>=1.0.1
certifi>=2024.2.2
rich>=13.7.1

# LLM and AI libraries
litellm>=1.74.12
posthog>=3.20.0

# Computer interface dependencies
pillow>=10.2.0
websocket-client>=1.8.0
websockets>=12.0
pyautogui>=0.9.54
pynput>=1.8.1
pyperclip>=1.9.0

# Computer Vision and ML (SOM)
torch>=2.2.1
torchvision>=0.17.1
ultralytics>=8.1.28
easyocr>=1.7.1
numpy>=1.26.4
opencv-python-headless>=4.11.0.86
matplotlib>=3.8.3
huggingface-hub>=0.21.4
supervision>=0.25.1

# Optional ML dependencies (for UITARS)
accelerate
transformers>=4.54.0

# Web server dependencies
fastapi>=0.111.0
uvicorn[standard]>=0.27.0

# UI dependencies
gradio>=5.23.3
datasets>=3.6.0

# CLI utilities
yaspin>=3.1.0

# MCP Server
mcp>=1.6.0,<2.0.0

# Platform-specific dependencies (macOS)
# Uncomment if running on macOS:
# pyobjc-framework-Cocoa>=10.1
# pyobjc-framework-Quartz>=10.1
# pyobjc-framework-ApplicationServices>=10.1
# mlx-vlm>=0.1.27

# Platform-specific dependencies (Linux)
# Uncomment if running on Linux:
# python-xlib>=0.33

# Platform-specific dependencies (Windows)
# Uncomment if running on Windows:
# pywin32>=310

# Development and testing dependencies
black>=23.0.0
ruff>=0.9.2
mypy>=1.10.0
pytest>=8.0.0
pytest-asyncio>=0.21.1
pytest-cov>=4.1.0
pytest-mock>=3.10.0
pytest-xdist>=3.6.1
ipykernel>=6.29.5
jupyter>=1.0.0
jedi>=0.19.2
types-requests>=2.31.0
aioresponses>=0.7.4
setuptools>=75.8.1

# HUD integration
hud-python==0.4.19

# Documentation (optional)
# mkdocs>=1.5.0
# mkdocs-material>=9.2.0
